requirements.txt
tensorflow
tensorflow-hub
imagehash
numpy
scipy
Pillow
scikit-image
jsonpickle
flask
pyzbar
opencv-python==4.5.4.60
keras_ocr